Implement Drag & Drop support and max thumbnail size option.
Fixes #3.

This commit implements Drag & Drop support in Budgie Desktop View. You can now copy files and symlink directories. For files and directories which already exist will, we will attempt append (Copy) to the end of the file name.

This functionality also introduces two new right-click options for FileItems in the FileMenu:

1. Cancel Copy - Cancels the copy operation
2. Move to Trash - Moves an item to Trash.

Added block cursor for when a file is being copied.