Fix failure to acquire mount or volume UUIDs. Fixes #11.
"Mounts" provided via file transfer or PTP options for mobile and MTP would fail to get an identifier for the mount or the mount's volume, causing us to segfault when these devices were plugged in.

We will now fall back to the drive's identifier. If we fail, we return an explicit FAILED_TO_GET_UUID so it's obvious to any use of get_mount_uuid that we should return early and not attempt to render it.