Add Nix support
This commit provides a Nix-based package definition and development shell. A Flake is also provided for environments where that feature is enabled.

To build the package derivation use `nix build` (with Flakes) or `nix-build -E "with import <nixpkgs> {}; callPackage ./. {}"` (without Flakes).

To enter the development shell, run `nix develop` (with Flakes) or `nix-shell` (without Flakes).

Signed-off-by: Federico Damián Schonborn <fdschonborn@gmail.com>