Make git rev-parse not fatal.
Some end users may opt to build this in an environment oriented towards source-based building where they have git installed but intend on using release tarballs (which themselves do not contain the git repo). While it is suggested to build in a clean environment such as a chroot, which is done in other OSes, fix this so they can at least unbork builds.

Fixes #16