drop all support code for creating config.h
Nowhere is a config.h used, so this is wasted effort and performs
unnecessary string shuffling in Meson, then pointless file I/O.

More impactfully, this adds an arbitrary dependency on building the
project from a git clone instead of a `meson dist` produced release
tarball, if git is installed. There are a couple possible approaches to
making this better, for example checking if `.git` exists in the root of
the repository before trying to run git. But it turns out to be totally
unnecessary, since the parsed information is never, ultimately, used. So
let's just drop this code entirely, instead of trying to make it run
more effectively.